refactor(server): split server implementation into separate file
